Add udev rule for the backlight HID device
This udev rule should match the backlight HID device and make it
available to logged in users. In addition to this it will create
symlinks like this:

/dev/asdbl-xxxxxxxx-xxxxxxxxxxxxxxxx

The x are placeholders for the serial number of the display. This is to
support setups with more then one Apple Studio Display connected at the
same time.